requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
pandas==1.5.3
numpy==1.24.3
openpyxl==3.0.9
flask-cors==3.0.10
xlsxwriter==3.0.3
sqlalchemy==1.4.46
pyodbc==4.0.35
python-dotenv==0.19.2
psycopg2-binary==2.9.5